Fix metric computation reset state issue
This is a major fix for using the hypermodel and a tuner. The reported metrics on the val/dev datasets contained information about the train set too as if there were no reset_state call, in turn the objective metric used by the tuner was garbage (the standalone computation of this metric would give very different results). This most likely explains the U shaped performance curves in time for different models (the model would pick up the metric information from a trained model and progressively introduce information about his own performance, introducing a slow decrease until the new model is trained enough to start re increasing the model performance)